+++ resources/lang/de.php
@@ -304,6 +304,9 @@ return [
                 "chloride" => [
                     "label" => "Chlorid",
                 ],
+                "managanese" => [
+                    "label" => "Mangan",
+                ],
             ],
         ]
     ]